chore(ISSUE-73): pylance 타입 오류 정리 + 잔재 감사 보고 (#74)
* chore(types): pyright 타입 불일치 정리

에디터(Pylance) 에러 제거 목적, 런타임 동작 불변.

- Mr2sSolverProtocol.evaluator 를 read-only property 로 선언 — mutable
  속성 선언이면 DnCMr2sSolver 의 property 구현이 구조적으로 불일치
- Edge.weight / AdjEntry.weight / get_indicator_function weight 를
  float 로 — harmonic super edge 가중치가 실수라 int 힌트는 거짓
- chain_contraction 의 Graph 생성을 dict 전달로 통일
- BinaryPolynomial 을 dimod.higherorder.polynomial 에서 직접 import

Refs #73


* docs: pair_key 삭제·멀티그래프 잔재 감사 보고 추가

적대적 리뷰 에이전트 감사 결과 기록: viz 스크립트 크래시 2건,
pair_key 잔존 코드, 죽은 함수·방어 코드 목록과 수정 우선순위.
ISSUE-73 후속 작업의 기준 문서.

Refs #73


* chore(types): 라이브러리 pyright 오류 0 으로 정리

pyright 기준 mr2s_module 의 71건 중 라이브러리 18건을 제거했다. 단순 무시가
아니라 실제로 틀린 annotation 을 바로잡는 쪽으로 처리했다.

- EmbeddableGraphPartition/DnCSolution 의 embedding_estimates 를
  list[EmbeddingEstimate | None] 로 수정. VertexCount 전략은 원래 [None] 을
  채우고 있었고, 이를 맞추려던 embedding_aware 의 None 필터는 sub_graphs 와
  인덱스 정렬을 깨뜨릴 수 있어 제거했다.
- DnCMr2sSolver.mr2s_solver 를 Mr2sSolverProtocol 로 되돌리고, build_bqm 을
  실제로 쓰는 임베딩 전략에 넘기는 지점에서만 QuboBackedMr2sSolverProtocol 을
  runtime 검사한다. SA inner solver(create_dnc_sa_solver) 가 다시 유효해진다.
- BinaryPolynomial/make_quadratic 을 dimod.higherorder 하위 모듈에서 직접
  import (reportPrivateImportUsage).
- MultiGraph 의 key dict 를 edge id 로 읽는 계약을 util/nx_multigraph.py 의
  multi_edge_copies 한 곳에 모으고, ILS·robbins 가 이를 쓴다.
- planar_layout 좌표 cast 를 planar_position_map 으로 단일화.
- NHopPolyGenerator 는 small_world_spec 미지정 시 명시적 ValueError.

전체 테스트 245 passed (slow 15 deselected).

Refs #73


* refactor(util): pair_key(정점쌍 키) 잔재 제거, 면 API 를 edge id 로 통일

정점쌍 frozenset 키 기반 면 API 는 평행 간선을 붕괴시키고, 프로덕션은 이미
edge-id 경로만 쓴다. 소비자였던 viz 스크립트 2개를 edge-id 모델로 옮기고
죽은 API 를 삭제했다.

삭제:
- planar_graph 의 EdgeKey / face_edges / build_face_edges_map (정점쌍 키 삼형제)
- planar_graph 의 clone_edge — 호출 0곳이면서 id 를 새로 발급해 쓰는 순간 버그
- solver/dnc_graph_partition_strategy.py — 순수 re-export, import 0곳

추가:
- inner_faces_by_edge_id / face_vertex_ring — edge subdivision 위에서 유계 면을
  edge-id step 으로 열거. 평행 간선이 살아남는다.

고친 버그:
- tests/cycle/partition_visualization: 면 색칠이 100% 죽어 있었다. 정점쌍 키
  집합과 edge id 집합을 교집합해 owner 가 항상 None (실측 patches=0). edge-id
  기준으로 고쳐 seed42 108면 / seed7 144면 전부 채색된다.
- tests/visualize_wall_protection: 실행 즉시 KeyError 로 크래시(정점쌍 키를
  edge-id API 에 투입). subdivision + edge-id 로 재작성하고 그래프 생성은
  공용 fixture(delaunay_graph_with_pos)로 통합했다. 실행 결과 외벽 보호 ON 은
  wall-damage=0, OFF 는 12 시드 전부 외벽 파괴(2~5) + macro 붕괴(예 8→2).

회귀 방지: test_draw_partition_fills_every_inner_face 가 폴리곤 수 == 내부 면
수를 검증한다 (PNG 크기만 보던 기존 테스트는 이 버그를 못 잡았다).

전체 테스트 246 passed.

Refs #73


* fix(solver): 병합 해의 strong_connect_rate 를 실측값으로 되돌림

DnC 는 Evaluator 가 병합 해에서 계산한 strong_connect_rate 를 버리고 자식
solution rate 의 곱으로 덮어쓰고 있었다. 전역 강연결과 다른 값이라 지표 의미가
오염됐고, 실제로 강연결이 아닌 병합 해에 1.0 을 보고하고 있었다.

실측 (delaunay n=60, keep=60%, w∈1..10, dnc_qubo_sa, 6 시드): 전 시드 병합 해가
강연결이 아닌데(true_global_sc=False) 옛 지표는 1.0000 / 1.0000 / 0.7 / 0.24 …
를 보고했다. 새 지표는 6/6 시드에서 실측 강연결과 일치한다.

근본 원인은 두 겹이었다:
1. merge_solutions 가 자식[0] 의 sample_set 을 병합 해에 물려줬다. 그 sample 의
   변수는 부모 간선을 일부만 덮어서(나머지는 _safe_lookup → 0 → 기본 정방향)
   Evaluator 가 merged_edges 가 아닌 엉뚱한 배향을 채점했다. → 빈 sample_set 으로
   바꿔 edges 기준 채점을 강제한다.
2. score_merged_solution 의 자식 rate 곱 덮어쓰기 → 삭제. Evaluator 실측값을 쓴다.

reduction lift 도 같은 결함이 있었다 (축약된 체인 간선은 inner sample 에 없다).
result.has_chains 이면 sample_set 을 비운다.

주의: DnC/reduction 해의 sample_score 는 이제 inf 다 (병합·lift 해에는 대응하는
에너지가 없다). score.strong_connect_rate 를 읽던 기존 벤치 수치는 이 커밋 기준으로
재측정해야 한다 — 옛 값은 전역 강연결을 과대보고했다.

테스트 262 passed (slow 15 포함).

Refs #73


* refactor: 죽은 방어 코드 정리와 중복 로직 단일화

감사 보고(#9·#10·#11·APSP-2) 항목을 처리했다.

- Evaluator._safe_lookup/_sample_to_directed_edges 가 qubo/solution_processing 의
  safe_lookup/process_solution 을 그대로 복제하고 있었다 (sample→방향 변환 두 벌).
  Evaluator 가 공용 코어에 위임하도록 바꿔 분기 위험을 없앴다.
- DnC._apply_merged_directions 의 `if edge is not None` 조용한 스킵 → KeyError.
  solution 과 graph 가 어긋나면 일부만 배향된 해가 정상인 척 흘러갔다.
  (reduction lift 는 이미 raise 였다 — 일관성 확보.)
- QuboSolver.run_with_embedding 이 _fixed_embedding_child_sampler() 본문을
  복붙하고 있던 것을 메서드 호출로 교체.
- target_k 이진 탐색 전략의 getattr/hasattr 춤을 TunableFaceCycleProtocol 로 대체.
  target_k 없는 face_cycle 이면 이진 탐색이 같은 run(graph) 를 반복할 뿐이라
  타입으로 요구를 명시한다 (vertex_count 의 도달 불가 기본값 2 도 제거).

테스트 262 passed, pyright(mr2s_module) 0 errors.

Refs #73


* chore(types): 테스트·시각화 pyright 오류 정리 (전체 0)

남아 있던 46건을 정리했다. 무시(ignore) 없이 실제 타입/버그를 고쳤다.

라이브러리:
- PositionMap 을 Mapping[Any, Point] 로 (Point = Sequence[float] | ndarray).
  Mapping 은 키가 불변이라 Hashable 로 묶으면 dict[int, ndarray] 가 안 들어온다.
- face_centroid 계산을 np.asarray(dtype=float) 로 명시.

테스트에서 드러난 실제 결함:
- visualize_macro_weak_dual: bipartite 분해가 AmbiguousSolution 으로 빠지면
  color_a/color_b 가 미정의 → `len(color_a)` 가 NameError 를 내고 bare except 가
  삼켜서 제목이 항상 "(0L | 0R bipartite)" 였다. except 분기에서 채우고 예외 삼킴 제거.
- test_qubo_mr2s_solver 의 StubEvaluator 가 Score 가 아니라 float 를 돌려주고
  테스트는 solution.score == 1.0 을 단언하고 있었다 (Score 계약 위반). Score 반환으로
  고치고 단언을 score.apsp_sum 으로 바꿨다.
- test_face_clusterer 의 boundary/repair 인자가 아직 정점쌍 튜플이었다 → edge id 로.

나머지는 스텁 한계 대응 (dimod/nx/matplotlib): multi_edge_copies·PositionMap 같은
공용 경계에 cast 를 한 번씩 두고, 테스트 대역은 cast 로 프로토콜 경계를 명시.
matplotlib 은 plt.Polygon → matplotlib.patches.Polygon, plt.cm.X.colors →
matplotlib.colormaps[...] 로 공개 API 사용.

pyright(mr2s_module + tests) 0 errors / 0 warnings, 테스트 262 passed.
viz 스크립트 2종 실행 확인.

Refs #73


---------

+++ mr2s_module/domain/adj_entry.py
@@ -3,6 +3,6 @@ from typing import NamedTuple
 
 class AdjEntry(NamedTuple):
   vertex: int # Could be destination or origin vertex
-  weight: int
+  weight: float
   directed: bool # if directed is true, then vertex is destination
   edge_id: int # 원본 Edge.id. QUBO 변수명(e_{edge_id}) 생성에 사용.
